requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-aiplatform>=1.70.0
google-cloud-firestore>=2.13.0
python-dotenv==1.0.0
